Function pallet_broker::pallet::dispatchables::enable_auto_renew

pub fn enable_auto_renew<T: Config>(
    core: CoreIndex,
    task: TaskId,
    workload_end_hint: Option<Timeslice>,
)

Extrinsic for enabling auto renewal.

Callable by the sovereign account of the task on the specified core. This account will be charged at the start of every bulk period for renewing core time.

origin: Must be the sovereign account of the task

core: The core to which the task to be renewed is currently assigned.

task: The task for which we want to enable auto renewal.

workload_end_hint: should be used when enabling auto-renewal for a core that is not expiring in the upcoming bulk period (e.g., due to holding a lease) since it would be inefficient to look up when the core expires to schedule the next renewal.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::enable_auto_renew.