Type Alias pallet_broker::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn current_timeslice() -> Timeslice

pub fn latest_timeslice_ready_to_commit(config: &ConfigRecordOf<T>) -> Timeslice

pub fn next_timeslice_to_commit( config: &ConfigRecordOf<T>, status: &StatusRecord ) -> Option<Timeslice>

pub fn account_id() -> T::AccountId

pub fn sale_price( sale: &SaleInfoRecordOf<T>, now: BlockNumberFor<T> ) -> BalanceOf<T>

impl<T: Config> Pallet<T>

pub fn configure( origin: OriginFor<T>, config: ConfigRecordOf<T> ) -> DispatchResultWithPostInfo

Configure the pallet.

origin: Must be Root or pass AdminOrigin.

config: The configuration for this pallet.

pub fn reserve( origin: OriginFor<T>, workload: Schedule ) -> DispatchResultWithPostInfo

Reserve a core for a workload.

origin: Must be Root or pass AdminOrigin.

workload: The workload which should be permanently placed on a core.

pub fn unreserve( origin: OriginFor<T>, item_index: u32 ) -> DispatchResultWithPostInfo

Cancel a reservation for a workload.

origin: Must be Root or pass AdminOrigin.

item_index: The index of the reservation. Usually this will also be the index of the core on which the reservation has been scheduled. However, it is possible that if other cores are reserved or unreserved in the same sale rotation that they won’t correspond, so it’s better to look up the core properly in the Reservations storage.

pub fn set_lease( origin: OriginFor<T>, task: TaskId, until: Timeslice ) -> DispatchResultWithPostInfo

Reserve a core for a single task workload for a limited period.

In the interlude and sale period where Bulk Coretime is sold for the period immediately after until, then the same workload may be renewed.

origin: Must be Root or pass AdminOrigin.

task: The workload which should be placed on a core.

until: The timeslice now earlier than which task should be placed as a workload on a core.

pub fn start_sales( origin: OriginFor<T>, initial_price: BalanceOf<T>, core_count: CoreIndex ) -> DispatchResultWithPostInfo

Begin the Bulk Coretime sales rotation.

origin: Must be Root or pass AdminOrigin.

initial_price: The price of Bulk Coretime in the first sale.

core_count: The number of cores which can be allocated.

pub fn purchase( origin: OriginFor<T>, price_limit: BalanceOf<T> ) -> DispatchResultWithPostInfo

Purchase Bulk Coretime in the ongoing Sale.

origin: Must be a Signed origin with at least enough funds to pay the current price of Bulk Coretime.

price_limit: An amount no more than which should be paid.

pub fn renew( origin: OriginFor<T>, core: CoreIndex ) -> DispatchResultWithPostInfo

Renew Bulk Coretime in the ongoing Sale or its prior Interlude Period.

origin: Must be a Signed origin with at least enough funds to pay the renewal price of the core.

core: The core which should be renewed.

pub fn transfer( origin: OriginFor<T>, region_id: RegionId, new_owner: T::AccountId ) -> DispatchResult

Transfer a Bulk Coretime Region to a new owner.

origin: Must be a Signed origin of the account which owns the Region region_id.

region_id: The Region whose ownership should change.

new_owner: The new owner for the Region.

pub fn partition( origin: OriginFor<T>, region_id: RegionId, pivot: Timeslice ) -> DispatchResult

Split a Bulk Coretime Region into two non-overlapping Regions at a particular time into the region.

origin: Must be a Signed origin of the account which owns the Region region_id.

region_id: The Region which should be partitioned into two non-overlapping Regions.

pivot: The offset in time into the Region at which to make the split.

pub fn interlace( origin: OriginFor<T>, region_id: RegionId, pivot: CoreMask ) -> DispatchResult

Split a Bulk Coretime Region into two wholly-overlapping Regions with complementary interlace masks which together make up the original Region’s interlace mask.

origin: Must be a Signed origin of the account which owns the Region region_id.

region_id: The Region which should become two interlaced Regions of incomplete regularity.

pivot: The interlace mask of one of the two new regions (the other is its partial complement).

pub fn assign( origin: OriginFor<T>, region_id: RegionId, task: TaskId, finality: Finality ) -> DispatchResultWithPostInfo

Assign a Bulk Coretime Region to a task.

origin: Must be a Signed origin of the account which owns the Region region_id.

region_id: The Region which should be assigned to the task.

task: The task to assign.

finality: Indication of whether this assignment is final (in which case it may be eligible for renewal) or provisional (in which case it may be manipulated and/or reassigned at a later stage).

pub fn pool( origin: OriginFor<T>, region_id: RegionId, payee: T::AccountId, finality: Finality ) -> DispatchResultWithPostInfo

Place a Bulk Coretime Region into the Instantaneous Coretime Pool.

origin: Must be a Signed origin of the account which owns the Region region_id.

region_id: The Region which should be assigned to the Pool.

payee: The account which is able to collect any revenue due for the usage of this Coretime.

pub fn claim_revenue( origin: OriginFor<T>, region_id: RegionId, max_timeslices: Timeslice ) -> DispatchResultWithPostInfo

Claim the revenue owed from inclusion in the Instantaneous Coretime Pool.

origin: Must be a Signed origin of the account which owns the Region region_id.

region_id: The Region which was assigned to the Pool.

max_timeslices: The maximum number of timeslices which should be processed. This may effect the weight of the call but should be ideally made equivalant to the length of the Region region_id. If it is less than this, then further dispatches will be required with the region_id which makes up any remainders of the region to be collected.

pub fn purchase_credit( origin: OriginFor<T>, amount: BalanceOf<T>, beneficiary: RelayAccountIdOf<T> ) -> DispatchResult

Purchase credit for use in the Instantaneous Coretime Pool.

origin: Must be a Signed origin able to pay at least amount.

amount: The amount of credit to purchase.

beneficiary: The account on the Relay-chain which controls the credit (generally this will be the collator’s hot wallet).

pub fn drop_region( _origin: OriginFor<T>, region_id: RegionId ) -> DispatchResultWithPostInfo

Drop an expired Region from the chain.

origin: Can be any kind of origin.

region_id: The Region which has expired.

pub fn drop_contribution( _origin: OriginFor<T>, region_id: RegionId ) -> DispatchResultWithPostInfo

Drop an expired Instantaneous Pool Contribution record from the chain.

origin: Can be any kind of origin.

region_id: The Region identifying the Pool Contribution which has expired.

pub fn drop_history( _origin: OriginFor<T>, when: Timeslice ) -> DispatchResultWithPostInfo

Drop an expired Instantaneous Pool History record from the chain.

origin: Can be any kind of origin.

region_id: The time of the Pool History record which has expired.

pub fn drop_renewal( _origin: OriginFor<T>, core: CoreIndex, when: Timeslice ) -> DispatchResultWithPostInfo

Drop an expired Allowed Renewal record from the chain.

origin: Can be any kind of origin.

core: The core to which the expired renewal refers.

when: The timeslice to which the expired renewal refers. This must have passed.

pub fn request_core_count( origin: OriginFor<T>, core_count: CoreIndex ) -> DispatchResult

Request a change to the number of cores available for scheduling work.

origin: Must be Root or pass AdminOrigin.

core_count: The desired number of cores to be made available.

pub fn notify_core_count( origin: OriginFor<T>, core_count: CoreIndex ) -> DispatchResult