Type Alias pallet_broker::ConfigRecordOf

pub type ConfigRecordOf<T> = ConfigRecord<BlockNumberFor<T>, RelayBlockNumberOf<T>>;

Aliased Type

struct ConfigRecordOf<T> {
    pub advance_notice: <<<T as Config>::Coretime as CoretimeInterface>::RelayChainBlockNumberProvider as BlockNumberProvider>::BlockNumber,
    pub interlude_length: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number,
    pub leadin_length: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number,
    pub region_length: u32,
    pub ideal_bulk_proportion: Perbill,
    pub limit_cores_offered: Option<u16>,
    pub renewal_bump: Perbill,
    pub contribution_timeout: u32,
}

Fields
advance_notice: <<<T as Config>::Coretime as CoretimeInterface>::RelayChainBlockNumberProvider as BlockNumberProvider>::BlockNumber

The number of Relay-chain blocks in advance which scheduling should be fixed and the Coretime::assign API used to inform the Relay-chain.

interlude_length: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number

The length in blocks of the Interlude Period for forthcoming sales.

leadin_length: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number

The length in blocks of the Leadin Period for forthcoming sales.

region_length: u32

The length in timeslices of Regions which are up for sale in forthcoming sales.

ideal_bulk_proportion: Perbill

The proportion of cores available for sale which should be sold.

If more cores are sold than this, then further sales will no longer be considered in determining the sellout price. In other words the sellout price will be the last price paid, without going over this limit.

limit_cores_offered: Option<u16>

An artificial limit to the number of cores which are allowed to be sold. If Some then no more cores will be sold than this.

renewal_bump: Perbill

The amount by which the renewal price increases each sale period.

contribution_timeout: u32

The duration by which rewards for contributions to the InstaPool must be collected.