Type Alias pallet_broker::OnDemandRevenueRecordOf

pub type OnDemandRevenueRecordOf<T> = OnDemandRevenueRecord<RelayBlockNumberOf<T>, RelayBalanceOf<T>>;

Aliased Type

struct OnDemandRevenueRecordOf<T> {
    pub until: <<<T as Config>::Coretime as CoretimeInterface>::RelayChainBlockNumberProvider as BlockNumberProvider>::BlockNumber,
    pub amount: <<T as Config>::Coretime as CoretimeInterface>::Balance,
}

Fields
until: <<<T as Config>::Coretime as CoretimeInterface>::RelayChainBlockNumberProvider as BlockNumberProvider>::BlockNumber

The height of the Relay-chain at the time the revenue request was made.

amount: <<T as Config>::Coretime as CoretimeInterface>::Balance

The accumulated balance of on demand sales made on the relay chain.